+++ src/Application/DTOs/DashboardConfigDto.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Hamzi\CoreWatch\Application\DTOs;
 
+use Hamzi\CoreWatch\Support\PackageVersion;
 use Hamzi\CoreWatch\Support\Translation;
 
 final readonly class DashboardConfigDto
@@ -30,7 +31,7 @@ final readonly class DashboardConfigDto
             ),
             'locale' => app()->getLocale(),
             'labels' => Translation::all(),
-            'version' => '2.1.0',
+            'version' => PackageVersion::current(),
         ];
     }
 }